some extra fixes and connection over web client
